+++ sentry-android-core/src/test/java/io/sentry/android/core/AndroidOptionsInitializerTest.kt
@@ -301,31 +301,6 @@ class AndroidOptionsInitializerTest {
         assertNull(actual)
     }
 
-    @Test
-    fun `init should set SdkVersion`() {
-        val sentryOptions = SentryAndroidOptions()
-        val mockContext = mock<Context>()
-        whenever(mockContext.applicationContext).thenReturn(null)
-
-        AndroidOptionsInitializer.init(sentryOptions, mockContext)
-
-        assertNotNull(sentryOptions.sdkVersion)
-        val sdkVersion = sentryOptions.sdkVersion!!
-
-        assertEquals(BuildConfig.SENTRY_CLIENT_NAME, sdkVersion.name)
-        assertEquals(BuildConfig.VERSION_NAME, sdkVersion.version)
-
-        assertTrue(sdkVersion.packages!!.any {
-            it.name == "maven:sentry-android-core"
-            it.version == BuildConfig.VERSION_NAME
-        })
-
-        assertTrue(sdkVersion.packages!!.any {
-            it.name == "maven:sentry-core"
-            it.version == BuildConfig.VERSION_NAME
-        })
-    }
-
     private fun createMockContext(): Context {
         val mockContext = ContextUtilsTest.createMockContext()
         whenever(mockContext.cacheDir).thenReturn(file)
